requirements.txt
asgiref==3.5.2
attrs==22.1.0
bleach==5.0.1
cffi==1.15.1
cryptography==38.0.1
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-ipware==4.0.2
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-api-logger==1.1.10
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.16.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
webencodings==0.5.1
